+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-Nothing here so far.
+### Fixed
+
+- `$jsinject` rules should not be included into regular content blockers: [#90]
+
+[#90]: https://github.com/AdguardTeam/SafariConverterLib/issues/90
 
 ## [v3.0.0]